etl_project/requirements.txt
lxml>=4.9.0
python-dateutil>=2.8.0
sqlalchemy>=2.0.0  # For database handling (useful for load_db.py)
python-dotenv>=1.0.0  # For loading .env files
flask>=2.0.0  # For a simple web server if needed for /web

